// Generador/csv-converter/src/dtos/rawV1.dtos.ts
// Generated by https://quicktype.io

export interface RawV1 {
  PDAT_Sample: number;
  SpeedMedian: number;
  SpeedAVG: number;
  DistancesAVG: number;
  TARGET_ID: number;
  SpeedMedian_target: number;
  Distance: number;
  Magnitude: number;
  Angle: number;
  Time: Date;
  Monotonic_ns: number;
  TDAT_speed?: number;
  TDAT_distance?: number;
}